+++ docs/examples/analysis.jl
@@ -1007,20 +1007,6 @@ cfr_prior_fig #hide
 #md # ```
 
 #md # ```@raw html
-#md # <details><summary>Submodel: confirmed_positivity_model</summary>
-#md # ```
-
-#md # ```@eval
-#md # using BVDOutbreakSize, CodeTracking, Markdown
-#md # Markdown.parse(string("```julia\n",
-#md #     (@code_string BVDOutbreakSize.confirmed_positivity_model(5)), "\n```"))
-#md # ```
-
-#md # ```@raw html
-#md # </details>
-#md # ```
-
-#md # ```@raw html
 #md # <details><summary>Submodel: test_sensitivity_model</summary>
 #md # ```
 
@@ -2091,8 +2077,7 @@ prior_chn = let
         export_case_days = obs.export_case_days,
         export_death_days = obs.export_death_days,
         breakpoint = breakpoint,
-        background_re = true,
-        confirmed_positivity_link = :composition,
+        background_pooling = background_pooling_model,
         genetic = genetic_seeding_model,
         tmrca_days = obs.tmrca_days
     )
@@ -3099,8 +3084,7 @@ pp_joint = predict(
             increments = missing,
         ),
         breakpoint = _BREAKPOINT,
-        background_re = true,
-        confirmed_positivity_link = :composition,
+        background_pooling = background_pooling_model,
         genetic = genetic_seeding_model,
         tmrca_days = obs.tmrca_days,
         ## The generator must be the model that was fitted. `n_patches`